moved game launching business out of GameConfig concerns
